Fix: resolve conflict
