+++ be/src/schedule/schedule.repository.ts
@@ -1,11 +1,6 @@
 import { DataSource, Repository } from "typeorm";
-import { AddScheduleDto } from "./dto/add-schedule.dto";
-import { Injectable, InternalServerErrorException } from "@nestjs/common";
-import { ulid } from "ulid";
+import { Injectable } from "@nestjs/common";
 import { ScheduleEntity } from "./entity/schedule.entity";
-import { HttpResponse } from "src/utils/http.response";
-import { ScheduleMetadataEntity } from "./entity/schedule-metadata.entity";
-import { UpdateScheduleDto } from "./dto/update-schedule.dto";
 import { DeleteScheduleDto } from "./dto/delete-schedule.dto";
 
 @Injectable()
@@ -13,63 +8,16 @@ export class ScheduleRepository extends Repository<ScheduleEntity> {
   constructor(dataSource: DataSource) {
     super(ScheduleEntity, dataSource.createEntityManager());
   }
-
-  async addSchedule(dto: AddScheduleDto, scheduleMetadata: ScheduleMetadataEntity) {
-    const { endAt } = dto;
-
-    const scheduleUuid = ulid();
-
-    const schedule = this.create({
-      scheduleUuid,
-      startAt: null,
-      endAt,
-      finished: false,
-      failed: false,
-      remindMemo: "",
-      last: true,
-      metadataId: scheduleMetadata.metadataId,
-    });
-
-    try {
-      await this.save(schedule);
-      const body: HttpResponse = {
-        message: "일정 추가 성공",
-        data: {
-          scheduleUuid: scheduleUuid,
-        },
-      };
-
-      return JSON.stringify(body);
-    } catch (error) {
-      console.log(error);
-      throw new InternalServerErrorException();
-    }
-  }
-
   async getMetadataIdByScheduleUuid(scheduleUuid: string): Promise<number> {
     const record = await this.findOne({ where: { scheduleUuid } });
     return record.metadataId;
-  }
-
-  async updateSchedule(dto: UpdateScheduleDto): Promise<void> {
-    const { scheduleUuid, startAt, endAt } = dto;
-    const record = await this.findOne({ where: { scheduleUuid } });
-    record.startAt = startAt;
-    record.endAt = endAt;
-
-    try {
-      await this.save(record);
-    } catch (error) {
-      console.log(error);
-      throw new InternalServerErrorException();
-    }
   }
 
   async deleteSchedule(dto: DeleteScheduleDto) {
     const { scheduleUuid } = dto;
 
     const record = await this.findOne({ where: { scheduleUuid } });
-    this.softDelete({ scheduleUuid });
+    await this.softDelete({ scheduleUuid });
 
     return record.metadataId;
   }
